// apps/web/src/api/generated/models/recommendationListItemDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Verdict API
 * OpenAPI documentation for the Verdict backend.
 * OpenAPI spec version: 1.0.0
 */
import type { RecommendationListItemDtoStatus } from './recommendationListItemDtoStatus';
import type { RecommendationPreviewMovieDto } from './recommendationPreviewMovieDto';

export interface RecommendationListItemDto {
  id: string;
  slug: string;
  status: RecommendationListItemDtoStatus;
  moods: string[];
  /** @nullable */
  groupType: string | null;
  /** @nullable */
  duration: string | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  description: string | null;
  createdAt: string;
  movies: RecommendationPreviewMovieDto[];
}